Moved everything into a package and added a package object where the Slack and Redis dependent variables are instantiated first, rather than in the bot object - this ensures that these variables are instantiated before any of the other classes are, since if a class with the dependent variable is instantiated first, the latter will be null
